rename: dialog 별 분리
